requirements.txt
faster-whisper>=1.0.0
PyAudioWPatch>=0.2.12
sounddevice>=0.4.6
numpy>=1.24.0
customtkinter>=5.2.0
pystray>=0.19.5
Pillow>=10.0.0
